Deleted all the files to rewrite with different tech
It was already a to-do-list app. It was designed with TailwindCSS.  But I wanted to design with MUI. Also, it didn't have update functionality which I wanted to add.  I wanted to add some more functionalities too. Moreover, I wanted to use Redux Toolkit. So, I planned to delete all the files and rewrite the whole application from scratch.